+++ tests/testthat/test-req-headers.R
@@ -20,18 +20,46 @@ test_that("can add repeated headers", {
   expect_equal(resp$headers$a, c("a,b"))
 })
 
+# redaction ---------------------------------------------------------------
+
 test_that("can control which headers to redact", {
-  expect_redact <- function(req, expected) {
-    expect_equal(attr(req$headers, "redact"), expected)
-  }
-
   req <- request("http://example.com")
-  expect_redact(req_headers(req, a = 1L, b = 2L), character())
-  expect_redact(req_headers(req, a = 1L, b = 2L, .redact = c("a", "b")), c("a", "b"))
-  expect_redact(req_headers(req, a = 1L, b = 2L, .redact = "a"), "a")
+  expect_redacted(req_headers(req, a = 1L, b = 2L), character())
+  expect_redacted(req_headers(req, a = 1L, b = 2L, .redact = "a"), "a")
+  expect_redacted(req_headers(req, a = 1L, b = 2L, .redact = c("a", "b")), c("a", "b"))
+})
 
-  expect_redact(req_headers_redacted(req, a = 1L, b = 2L), c("a", "b"))
+test_that("only redacts supplied headers", {
+  req <- request("http://example.com")
+  expect_redacted(req_headers(req, a = 1L, b = 2L, .redact = "d"), character())
+})
 
+test_that("redaction preserved across calls", {
+  req <- request("http://example.com")
+  req <- req_headers(req, a = 1L, .redact = "a")
+  req <- req_headers(req, a = 2)
+  expect_redacted(req, "a")
+})
+
+test_that("req_headers_redacted redacts all headers", {
+  req <- request("http://example.com")
+  expect_redacted(req_headers_redacted(req, a = 1L, b = 2L), c("a", "b"))
+})
+
+test_that("is case insensitive", {
+  req <- request("http://example.com")
+  req <- req_headers(req, a = 1L, .redact = "A")
+  expect_redacted(req, "A")
+  expect_snapshot(req)
+})
+
+test_that("authorization is always redacted", {
+  req <- request("http://example.com")
+  expect_redacted(req_headers(req, Authorization = "X"), "Authorization")
+})
+
+test_that("checks input types", {
+  req <- request("http://example.com")
   expect_snapshot(error = TRUE, {
     req_headers(req, a = 1L, b = 2L, .redact = 1L)
   })